fix: 내부 api 컨트롤러 분리
